+++ GPIO_ex.h
@@ -24,19 +24,19 @@
 ** \brief GPIO input structure
 **/
 typedef struct GPIO_in {
-	bool			in;						//!< Input value
-	eEdge			edge;					//!< Input edge
+	bool			in;					//!< Input value
+	eEdge			edge;				//!< Input edge
 	/*pvt*/
-	bool			mem;					//!< Memo value
-	uint32_t		hIn;					//!< Filter time
+	bool			mem;				//!< Memo value
+	uint32_t		hIn;				//!< Filter time
 	struct {
-	GPIO_TypeDef *	GPIOx;					//!< HAL GPIO instance
-	uint16_t		GPIO_Pin;				//!< HAL GPIO pin
-	uint16_t		filt;					//!< Filter time (ms)
-	bool			logic;					//!< Input logic polarity
-	bool			repeat;					//!< Callback ON repeat
-	void			(*onSet)(void*);		//!< Push callback ON function pointer
-	void			(*onReset)(void*);	//!< Push callback OFF function pointer
+	GPIO_TypeDef *	GPIOx;				//!< HAL GPIO instance
+	uint16_t		GPIO_Pin;			//!< HAL GPIO pin
+	uint16_t		filt;				//!< Filter time (ms)
+	bool			logic;				//!< Input logic polarity
+	bool			repeat;				//!< Callback ON repeat
+	void			(*onSet)(void *);	//!< Push callback ON function pointer
+	void			(*onReset)(void *);	//!< Push callback OFF function pointer
 	} cfg;
 } GPIO_in;
 
@@ -57,34 +57,36 @@ typedef struct GPIO_in {
 **/
 void NONNULLX__(1, 2) GPIO_in_init (GPIO_in * const in,
 									GPIO_TypeDef * const GPIOx, const uint16_t GPIO_Pin, const bool logic, const uint16_t filter,
-									void (*onSet)(GPIO_in*), void (*onReset)(GPIO_in*), const bool repeat);
+									void (*onSet)(GPIO_in * const), void (*onReset)(GPIO_in * const), const bool repeat);
 
 
 /*!\brief Handles GPIO_in read and treatment
 ** \param[in,out] in - input instance to handle
 ** \return Nothing
 **/
-void NONNULL__ GPIO_in_handler(GPIO_in * in);
+void NONNULL__ GPIO_in_handler(GPIO_in * const in);
 
 
 /*!\brief Get GPIO_in input value
 ** \param[in] in - input instance
 ** \return Input value
 **/
-__INLINE bool NONNULL_INLINE__ get_GPIO_in(const GPIO_in * const in) { return in->in; }
+__INLINE bool NONNULL_INLINE__ get_GPIO_in(const GPIO_in * const in) {
+	return in->in; }
 
 
 /*!\brief Get GPIO_in input edge
 ** \param[in] in - input instance
 ** \return Input edge
 **/
-__INLINE bool NONNULL_INLINE__ get_GPIO_in_edge(const GPIO_in * const in) { return in->edge; }
+__INLINE bool NONNULL_INLINE__ get_GPIO_in_edge(const GPIO_in * const in) {
+	return in->edge; }
 
 
 /*!\brief Get name from Port, Pin
 ** \param[in,out] name - pointer to string for name
-** \param[in] GPIOx - port to write to
-** \param[in] GPIO_Pin - pin to write to
+** \param[in] GPIOx - port
+** \param[in] GPIO_Pin - pin
 ** \return Error code
 **/
 FctERR NONNULL__ str_GPIO_name(char * name, const GPIO_TypeDef * const GPIOx, const uint16_t GPIO_Pin);
@@ -110,7 +112,7 @@ __INLINE void NONNULL_INLINE__ write_GPIO(GPIO_TypeDef * const GPIOx, const uint
 		#if defined(VERBOSE)
 			char port[10] = "";
 			str_GPIO_name(port, GPIOx, GPIO_Pin);
-			printf("Written %s to %u (%lums)\r\n", port, HAL_GPIO_ReadPin(GPIOx, GPIO_Pin), HAL_GetTick());
+			printf("Written %s to %u (%lums)\r\n", port, HAL_GPIO_ReadPin(GPIOx, GPIO_Pin), HALTicks());
 		#endif
 	}
 }
@@ -126,15 +128,15 @@ __INLINE GPIO_PinState NONNULL_INLINE__ read_GPIO(GPIO_TypeDef * const GPIOx, co
 	/* Check the parameters */
 	assert_param(IS_GPIO_PIN(GPIO_Pin));
 
+	const GPIO_PinState pin = HAL_GPIO_ReadPin(GPIOx, GPIO_Pin);
+
 	#if defined(VERBOSE)
-		GPIO_PinState pin = HAL_GPIO_ReadPin(GPIOx, GPIO_Pin);
 		char port[10] = "";
 		str_GPIO_name(port, GPIOx, GPIO_Pin);
-		printf("Read %s is %u (%lums)\r\n", port, pin, HAL_GetTick());
-		return pin;
-	#else
-		return HAL_GPIO_ReadPin(GPIOx, GPIO_Pin);
+		printf("Read %s is %u (%lums)\r\n", port, pin, HALTicks());
 	#endif
+
+	return pin;
 }